feat(lvar): add initial implementation LVar
I have little confidence in the interpreter implementation. I think what
needs to happen is that LInt and LVar have to have their interpreters
updated to accept a type parameter for the lang that is being parsed.
That way we don't have to rely on unsafeCoerce to get the type checker
to be happy.